requirements.txt
Flask
pyserial
waitress
